[penge] Add PengeAppsPane that shows a grid of PengeAppTile
PengeAppsPane takes the PengeAppBookmarkManager and then creates a
PengeAppTile for every bookmark in the manager layed out within a table.

PengeAppTile is used to give a graphical representation of a bookmark.

This is just an early version and does not yet handle additions or removals of
bookmarks nor is the tile interactive or styled correctly.