[status] Use Mojito to retrieve the last status
When a StatusRow has been constructed with a valid service name, we
create a MojitoClient and open a view on it. The view will return the
first status, inside a MojitoItem. The properties on the item are
used to set the status text on the MnbStatusEntry created by the row.