[status] Allow updating the status message
Clean up the output of the StatusEntry widget by splitting up the
status text entry from the status time and service label.

Added the StatusEntry::status-changed signal which returns the new
status text.

The ::status-changed signal is used by the StatusRow to call the
MojitoClientService API to update the status text.