[status] Refactor StatusEntry:active
The logic for toggling the :active state in the StatusEntry is duplicated
all over the place, and the only place where this should really be
present, mnb_status_entry_set_is_active(), does not work correctly.

The mnb_status_entry_set_is_active() method should be the only place where
the :active state is toggled, and the ::clicked signal handler of the
StatusEntry button should just toggle the :active state.