[MnbPanel] Connect netpanel to toolbar show/hide
Create the internet panel when showing the toolbar and destroy it when
hiding the toolbar, as opposed to when clicking on the internet panel
button. This is needed to avoid seeing the lag between connecting to a
view on the browser and getting the first update. This code may be
changed to work in idle, or something, to avoid possible lag when the
browser window is large or has multiple tabs open (requires testing).