[pasteboard] Add an expiration policy for ClipboardStore
We don't want the pasteboard to grow indefinitely, so the
ClipboardStore list model should apply an expiration policy.

The eviction is performend each time the contents of the
pasteboard change (i.e. something has been added) instead
of periodically. The expiration is performed based on the
time an item was added to the ClipboardStore.