[pasteboard] Rework to allow item removal
Various incremental changes to allow removing items from the Store
using the View, and also to get notification.

First of all, we add an ::item-removed signal to the Store which wraps
the ::row-removed signal coming from ClutterModel; this is needed
so that the View does not need to know the Store columns layout.

Then the View will connect to both ::item-added and ::item-removed
when it has been associated with a Store.

We also need the declaration of the get_type() function for the
Item class, and specify the style class name for the Action button
inside the Item.