[pasteboard] Remove the items after the store has changed
We should ask the store to remove an item using its serial, and
once the store complies and we get removal notification, we can
safely remove the Item widget from the View.

This also allows us to clear the View by simply iterating over
the Store and calling clutter_model_remove().