[theme] Make the MnbEntry style adapt to the page
Both the applications page and the pasteboard page use MnbEntry, which
means that the MnbEntry should provide a default style and the various
pages using it should just tweak it to match the indications of the
UX designs.

In this case, the difference is just the color of the button's text,
which matches the overall tone of the page.