[MnbNetPanel] Set Y-expand/fill as a work-around
NbtkTable was squashing us, despite x-fill/x-expand due to some
allocation funniness? Not sure of the reason, but this is a suitable
work-around for now.

Also add ! important to border-image: none for the button inside the
entry. Note, the display is still wrong after hovering over it, but
we'll be overhauling this panel at some point soon anyway.