[penge] Fix recent file tile web site behaviour
When we get URL that doesn't begin with file:// don't use g_file_query_info on
it since that tries to connect to connect to the internet via gvfs to fetch
the web page to determine that uhm...it's a web site. So this patch fixes that
and just identifies URLs starting with http:// as web pages and moves on.