[launcher] Prevent expanders' logical and css state go out of sync
When a launcher button was hovered in filter mode all widgets on the
grid were reset to css pseudo class NULL. If there was an invisible
opended expander it would rendered in an invalid state ("expanded" +
css pseudo class NULL).

The correct thing to do is only resetting the other launcher buttons
on the grid and skip the invisible expanders.