[Startup Notification] Fix the hashing of SN ids
When using GAppInfo, there is no way to obtain the SN id of the application we
just launched. We work around this by initially storing the startup data using
the application binary as a key, and once we get the SN INITIATED event, we
rekey the data with the sn_id.