[MnbDropDown, MnbToolbar] Reduce allocations
Do not do full reallocation when (a) we are not visible; (b) we are animating.
In the former case just return; this defers allocation till show(). In the
latter case we skip over the parent allocate directly to
ClutterActor::allocate(); this ensures our box is at the correct position (NB:
this is currently not foolproof, as we could end up with incorrect allocation
if allocation is queued up during the animation; in clutter 1.0 we will be able
to handle this correctly by being able to differentiate between allocations
happening because of absolute origin changing only, and those that were queued
up.)