[netpanel] Add default tab and favorite pages section placeholder
When there are no tabs open in the browser, or DBus communication fails, etc,
now the netpanel provides a default tab button that will open a new tab in
the browser. Also a stub favorite pages section is now there with a message
about how to theoretically populate it. Still need a default image for the
tab and to implement the favorites gathering.