[MnbToolbar, MnbPanel] Replace RequestIcon with RequestButtonStyle
Because of the way Nbtk styling works, we cannot just change the icon,
but need to change the style. As there is no way to set a style from
a string, each panel will need to provide an external stylesheet to be
used for it's button, and then request change of style when it needs to
change the icon.