[MnbPanel,MnbToolbar] Reworked handling of remote panel dying
Moving over the async initialization of the MnbPanel object introduced race
condtions into the MnbPanel code that was trying to restart the service. So,
instead of attempting to handle this in the MnbPanel object we emit
MnbPanel::remote-process-died signal, which we handle in the Toolbar
(remove the old panel and associated button, then attempt to restart the
service).