[App launching API] Fix confusion about parameter order
Through a fsck of mine it came to be that the chooser and workspace paramete
were in different order in the internal launching API exposed by the chooser
and the API exposed through the Toolbar dbus service. This was further
aggravated by the internal call made by the Toolbar passing the two parameters
in the wrong order, and masked by the fact that in some places, like the app
launcher, the order in the calls to the Toolbar API was also wrong. The current
commit fixes this by changing the internal API to use the same order as the
Toolbar API (gint workspace, gboolean no_chooser) and fixes up all the calls
both the internal and Toolbar API to pass parameters in that order.