--enable-cache configure option; disabled by default
Unfortunately, the current version of nbtk-create-image-cache cannot be run
during package build due to lack of differentiation between build root and
canonical install location (see bug 3977). rpm will detect the presence of the
build root string in the cache file and abort, so without this option we are
unable to package at all. This is intented as a temporary workaround until
the nbtk-create-image-cache tool is improved.