[AnerleyEContactItem] Support avatars on EContacts
There are two types of avatar available on EContacts, a URI based on and an
inline based one. We support the URI based one for local URIs only. For the
inline version we pull the content out of the vcard and save it to a file in
the avatars directory named after the uid.

This fixes MB:#3629