[pasteboard] Hide the action button for the first item
The first item in the pasteboard should not have the 'Copy' button
since the contents are already the one in the CLIPBOARD selection.

When we remove the first row, though, the newly moved up row is not
going to be in the CLIPBOARD, and thus needs to have the 'Copy'
button available.