[penge] Let the height of the welcome text determine the box it's in
Previously the welcome tile (shown on first 5 boots / if there are fewer than
recent files) had a fixed height. This doesn't work if you have a wordy
language. This change also includes a work around for an NbtkTable bug where
having a col-span of greater than the number of columns.

Fixes MB:#4687