CarrickStatusIcon: Workaround fix for delayed setting of status icon
Moblin bugzilla #4163
When the icon first draws itself the CmService object may not be fully
initialised such that we cannot tell what its type is and we draw an
error icon. In this case connect to the "service-updated" signal of the
CmService object and when we receive that re-draw the icon.