Add the destkop_tex to stage, rather than a clone
The base texture we are cloning needs to be always inserted into stage, so
if we insert the original as the background, and use a clone for the desktop
window. This way things work even when we have no desktop window, or the
the window mapped before we were started.