[moblin-panel] Delay bookmark removal if desktop file disappears
When a bookmarked application is upgraded thru the package manager
its desktop file is replaced. This triggers a notification so the
bookmark manager thinks the app had been uninstalled and removes
the bookmark.

Now, when a bookmarked desktop file disappears, a "removed bookmark"
file containing the desktop file uri is written. After a timeout of
10s the bookmark manager checks for the existence of the desktop file,
and if it is still gone the bookmark is removed.

When the bookmark manager initially loads the bookmarks it checks for
leftover removed bookmarks, in case the system had been shut down
during the timeout.

Fixes: MB:#4729