Don't overlay desktop window with our custom background
Until now, if a desktop window was present, we covered it up with the moblin
background texture. This was a left over from the early parallax effect, but
prevents a conventional desktop applications from working (or rather they work,
but their content is not visible, which is confusing). Since the parallax
is no longer requried, this is no longer necessary.