[penge] Adapt col-span workaround to take into consideration excluded files
In order to set the col-span correctly for the welcome tile on the Myzone we
need to take into consideration the number of tiles that are actually included
in the recent files pane. Not the number of items returned by the
GtkRecentManager. This is because the pane excludes files that have no
thumbnail or are missing from the filesystem.

This bug fixes: MB:#4991.