[netpanel] Perform a search for text that is clearly not a valid URL
Fixes BZ 2899. Added one hack to escape % signs as %% so that g_app_info
calls in the toolbar code will not break URL escaping like %20 for space.
Also noted a race condition in this code that will need to be fixed; if
the user types quickly and hits enter the search text will be incomplete.