[Zone switch effect] Don't raise the effect above the overlay
Remove some code left over from an earlier version of the effect when it was
inside the window group. Now that it is directly on stage, its depth is fixed
between the window_group and the overlay, so we do not need to mess about with
the stacking any longer.