Panel autostarting
We generally expect the panels to be pre-started for us in the session script,
to speed up the boot process, and we want to avoid falling back on the single
instance nature of the panels during the boot for autostarting to avoid loading
second instance of a panel even for a short period of time. Not autostarting,
however, comes with drawbacks; notably, if the WM is restarted we can end up
with no panels.

This commit implements delayed auto starting -- after a pre-defined period of
time elapses (currently 10s), we check for panels that are not running, and
attempt to start them.