[libmoblin-panel] Load the stylesheet that is common for all panels
This stylesheet contains styling for the MplEntry mainly. It is loaded when
the MplPanelClutter is constructed.

It also loads into the NbtkTextureCache a cache for *just* the assets that are
used by all panels.

A function is exposed, 'mpl_panel_clutter_load_base_style', that can be used
to load the style in standalone testing of panels.