Coding style: Use consistent method for accessing private structure
Use a consistent method for accessing the private member structure of the
object.
Convert all classes to the patter whereby the private structure is typedef'd
in the header and a pointer to it declared in a public (in the header) copy
of the private structure definition.
This allows usage of the object->priv access patter.