[MplPanelClutter] Pull initialization code from macros into functions
The initialization code was originally implemented as macros as we could not
use gtk/gdk code in the clutter version of the library; this is no longer the
case, since we now ship only one library.

This commit also ensures that we update the user time on our top level window
when appropriate.