[MnbSwitcher] Force hiding of visible tooltip when hiding the Switcher
This is only bullet proofing, since the tooltip should be getting destroyed
as soon as we destroy the switcher contents. But in case the item owning the
tooltip gets leaked (like when a bug in NbtkDraggable left an unbreakable
reference on the switcher items), we at least want to avoid the tooltip left
on screen for ever.