CarrickServiceItem: Show passphrases by default
This is a controversial change but only because everyone else does it the
other way so it has become what users expect.
The amount of incorrectly typed passphrase errors can surely be decreased
by this change.

Besides, my manager and UX designer both want it this way. And so do I! :-p