[MnbToolbar] Fix for Toolbar hiding (MB#3949)
We used to block the Toolbar hiding when The toolbar was already animating;
the consequence of this was that if the pointer moved fast enough during the
transition to leave the toolbar before it got completely shown, the toolbar
would remain visible, as no furether leave events would take place until the
pointer re-entered the toolbar. This commit disables this logic.