Workspace management fix
When checking whether workspace should be deleted or not, we need to
differentiate between the reason behind the check. If the check was triggered
by window moving onto a different workspace, we need to exlude transients of
the moved window from the check (since the WM will move these onto the
application workspace); however, if the check is triggered by  window being
destroyed, we need to ingore it's transients (and respond to the transient
itself being destroyed).