[MnbToolbar] Finegrained waiting_for_panel (MB#6402)
We need to differentiate between waiting for panel to show and waiting for
panel to hide; the buttons need to ignore clicks in both cases, but the leave
events should only be ignored if we are waiting for the panel to show. Also,
if we are waiting for a panel to hide, mnb_toolbar_panels_showing() needs to
return FALSE even if the panel is still mapped.