Special handling for panel children (MB#5499)
When a panel is visible, check whether any window that mapped belongs to the
panel (we do this by comparing the WM_CLASS property for both). If the window
belongs to a panel, we treat it in way similar to the scim window, by stacking
clones of the actual windows above the panel.