Only allow the short Alt+Tab when compositor is disabled (MB#6861)
When compositor is disabled, we have no way of providing visual feedback on
the tabbing, so I have disabled the long Alt+Tab press (which normally brings
down the compositor). The short Alt+Tab press, which switches immediately to
the next application on the tab list, is still available.

(This whole thing needs more though; it might be better to treat all alt+tabs
in this situation as a single short alt+tab.)