anerley-tp-feed: Listen for status changing on the connection
After removing the account from the account manager the signal saying that the
connection has changed state doesn't get relayed through to the TpAccount from
the TpConnection. A bug has been filed at
http://bugs.freedesktop.org/show_bug.cgi?id=25149

This change works around this issue by listening for the status property on
the connection changing as well as the TpAccount.

Fixes MB#8062