[MnbPanel->MnbPanelOop] Rework the panel object so that it is not an actor
As we are moving the window manipulation into the panel library, the panel
object becomes just a convenience wrapper unifying the MutterWindow of the
panel, and the client dbus interface. Renamed this panel object in
preparation for creating a MnbPanel interface that will be used by the
different panel implemenations.