[MnbToolbar, MnbPanel] Store input region per-panel, rather than globally
This is necessary now that the show and hide signals are getting passed over
dbus, we need to make sure that we are always removing the correct region for
the given panel. The implementation details are hidden in the input manager.

[MnbInputManager] Convenience API for automatically handling regular actors

debug spew

Separate input manager API for oop panel

remove debug spew