[MnbPanelOop] Change when we only hide panel, and when we also hide toolbar
This changes the default behaviour -- by default, when hiding the panel, we
also hide the toolbar; this way if the OOP process hides the panel, e.g.,
because it launched an application, we do the right thing.

Internally, all panel hiding is routed through the MnbPanel vfuction, and as
a rule, anything that uses the internal API will lead only to closing the
panel, not the toolbar (e.g., toggling toolbar buttons). If we want to close
both internally, we have use the mnb_panel_hide_with_toolbar() API.