[MplPanelClient] Added a "ready" signal
If the object is newly constructed, we emit the signal from an idle, ensuring
that the construction of Panel is completed and the main loop is ready to
spin. If the object is getting re-embedded, we emit the signal as soon as the
hand-shake is completed.

Some of the panels use idles to actually populate the panel contents; we allow
the main loop to spin few times before signalling we are ready and use low
priority idle as well.

For panels that do more complex set up, the automatic idle based emission is
not sufficient, and the signal emission needs to be postponed. For this
scenario the panel can be constructed with the delayed-ready property set to
TRUE, in which case the "ready" signal will not be emitted until
mpl_panel_client_ready() is called by the application.