[MnbToolbar] Fixed show animation completed callback
If the show animation is superceeded by a hide animation (i.e., the toolbar
get's hidden before it is completely shown) the show callback still gets
called. We must not show things like the shadow, etc, when this happens.

(This is probably a bug, need to investage further when/why this happens, but
if it does, we do not want to leave things in an inconsistent state.)