[constraints] implement moblin constraints in plugin
Implements the new constrain_window() vfunction of MutterPlugin (this replaces
previous constraints patches to Mutter itself).

Two constraints are implemented in this commit:

not_too_small constraint: windows over certain percentage of the screen are
enlarged and centered; this constraint is only applied in netbook mode.

screen_sized constraint: this constraint is applied to applications that set
moblin-frame-style hint to 'naked'; in netbook mode, such applications are
resized to fit the entire screen.