[MnbSwitcherApp] pass window directly into  hide-completed callback, MB#9244
At some point this code was modified to only activate the given application
after the Switcher is hidden; as the Switcher contents get destroyed on hide,
the MnbSwitcherApp instance passed into the hide-completed callback is no
longer valid, and this would cause irregular crashes. Simply pass the
MutterWindow into the callback instead.