[MnbToolbar] Avoid enlessly trying to load a failed panel
If panel fails to load in the given time, mark it as broken and do not try to
show it again (we change the tooltip to indicate the problem to user, perhaps
something bit more informative is needed instead).

This commit also reduces the time we wait for the panel to 6s.