[alttab] Handle kbd grab failure on initial Alt+Tab event, MB#9735
If the user hits the Alt+Tab just before the application grabs keyboard, we
fail to get a grab for ourselves, which means, we will not receive any further
kbd events, and the overlay just remains there until the users presses the
Alt+Tab again after the application released the kbd grab.

We can't really do the Alt+Tabbing withough having a grab, so if the grab fails
we simply do not show the overlay; this is not ideal, but is better than
leaving the overaly haning on the screen, and should be relatively rare.