Handle case when panel shows bypassing the map effect
There are certain circumstance in which the map() effect is skipped by the
compositor; this includes the case when a workspace switch effect is in
progress. Consequently, we cannot rely on showing our panels in the map()
effect.

The sequence of signals when the map effect *is* runs is:

  MetaDisplay::window-created -> map() -> ClutterActor::show(mcw)

So, when a panel window is created, we connect to the show signal, and then
in the map() vfunction disconnect the show callback; this way if the show
callback triggers, we know that the map() effect was skipped, and we let it
to show the panel instead of doing so in the map effect.