panel-status: Adjust the height of the panel so that is 600 pixels
This means that the panel will be basically full height on a netbook (shrunk a
bit) and then a fixed height on the nettop. We do this by making a height
request rather than tracking the child since then it adapts correctly.

This should fix http://bugzilla.meego.com/show_bug.cgi?id=520