events-pane: Correctly filter events
Since we now have no visual design for events in the past and because events
are not fixed sizes we cannot use the sliding window algorithm. We now simply
filter events out that have an end in the past.

This should fix: http://bugzilla.meego.com/show_bug.cgi?id=544