launcher-button: Use notify:toggled for pin state change
Previously the code listened for the clicked signal then checked for the
toggled state. However the internal toggled state wasn't set at this point in
under the new MX API since that is done in a default implementation. As an
alternative to using connect after on "clicked" listening for
"notify::toggled" is slightly more elegant.

Fixes: http://bugzilla.meego.com/show_bug.cgi?id=133