[libmoblin-panel] MplPanelBackground, a widget for Clutter panels
MplPanelBackground paints the panel background using the same asset as used for
the panel frame by mutter-moblin. The widget is automatically added to the
bottom of MplPanelClutter stage at object construction time, and requires no
more from the individual panels than to ensure that their top-level widget is
transparent.

MnbPanelFrame and MplPanelBackground styles need to be kept identical in
regards to border-image and padding.