Support a 'meego-no-dismiss' hint to hide 'Dismiss'
Adapted from patch by Aaron Bockover <abockover@novell.com>

In some cases, notifications do not want the default 'Dismiss' action,
so I add support for a 'no-dismiss' hint that will not show the default
'Dismiss' action.

In SUSE MeeGo we use this to alert the user that their system is about
to be registered for updates, but provide an opt-out window. The message
is something like this:

  *Register SUSE MeeGo*
  Your system is about to be registered
  for online software updates.
  [Register] [Do Not Register]

In other notification daemon implementations, instead of a "Dismiss"
action in the action area, a close button is added always on the
top-right side of the notification, so it's non-intrusive.