app-launches-store: Add asynchronous write
Asynchronous write is implemented by the "meego-app-launches-store"
binary, which is installed to $libexecdir. This binary is spawned
when async write is requested.

Also add the possibility to pass a timestamp when adding an app launch
to the database. This is needed to ensure correct launch time in case
async write is delayed by a lock on the database.

Update mpl-app-launch-context to use the async write API.