Import OpenConnect authdialog
This patch imports a copy of what used to be OpenConnect auth-dialog.
There's no point trying to share it any more; abusing the
NetworkManager GUI was a dirty hack from the beginning, and current
versions of NetworkManager don't even use the same GConf storage for
their configuration any more.

After this we are free to do whatever we like with it — change the
storage to something saner than GConf, integrate it properly into the
panel etc.

We may still want to keep it in a separate process, however, since it
needs to link against OpenSSL and hence is licensed under LGPL instead
of GPL.

Patch from David Woodhouse.

Fixes BMC#8114